#9 :  adjust diagram on interactions , refactor interactions lib to be more compliant with diagram
