+++ Vonage-Challenge.drawio
@@ -1,1 +1,1 @@
-<mxfile host="Electron" modified="2021-08-31T20:16:25.890Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/14.9.6 Chrome/89.0.4389.128 Electron/12.0.16 Safari/537.36" etag="NoeFmH7QGQsaVoNhoass" version="14.9.6" type="device"><diagram id="9mXvaVSuT-yr561JVfZQ" name="Page-1">7V1dc5s4FP01nkkf0jFgE+cxdpK2u8lutmnTdl92ZCxjNTLyguzY/fUrgTBgLg5JEE63emp0ETbonHu4H8LtOKP5+l2IFrNrNsG0Y3cn645z3rFtqztwxT/Sskksg64y+CGZqEmZ4Zb8wOmZyrokExwVJnLGKCeLotFjQYA9XrChMGQPxWlTRovfukA+LhluPUTL1i9kwmfqLvrdzP4eE3+WfrPVVUfmKJ2sDNEMTdhDzuRcdJxRyBhP/pqvR5jKxUvXJTnvsuLo9sJCHPA6J9xNx1cr8v3q8xd3+Q/6bfD565/r43SZI75J7xhPxAKoIQv5jPksQPQisw5DtgwmWH5sV4yyOVeMLYTREsbvmPONQhMtOROmGZ9TdTTiKORnEh1hCFiAU9sloVR9LA4m6YwxZd59YspNSC5cXm3lgihTxJahp2aFd/Ry/vfJh03v9w+X795P7NuL2bGtiIVCH/N9q7WFTfAdsznm4UacF2KKOFkVrwMp4vnbeRk24g8FDwzVvotcIbpU33SNo0jS1+5+CDgOkccJC0p4Rg9kTlG8xlMW8BRaiQOixA/E3x6W5wvDCoecCPKfqQNcAjr0ZoROrtCGLeXiCKC8+3Q0nLGQ/BAfiwrYKuRttzDjVp6p0AtxJObcpGBZO6ZrtC5MvEIRVwaPUYoWERlvb2MucCPBkHHO5gVyAFyQd4jXOVMZTnXUTn1cKdepqxB4yHTActWcWU4DnJ4mBvRKDCiBTYlyppDdbwVLrtFUOM6IURZmHpeiT/GUA9jPyWQSu3u0QB4J/E+xcx9bmeUqPvHcySwf1So4sUpwxFGCkoSEojGmNywiMUud8zCZO1wwEvB4qfrDTv88toR8xAJxE4jE8GGB/gOWDACA3essj6Ot0LXdeuAOXo4tLC1WCctfUIjdn0GI3ZIb3szk0hkZ1iHD1mtT4YFR4boq7NYGu30VBi/YKgdZZ768R+PNzXjzYMedB063JuKuLsidEuTHEuvlAocrEklXPRPj22y8ywVx63zr7Tve/XSHT9hR9O1e2beliYlzpzR+Ms+ETOAA8PeiXw/Fyo66b/vSw+2RGFvZuCmn3/pQ415va+OAiaw1wtuv+cTWJup9AN2hrC1hnrn10ZuCoxsvb5oGbk2lb8LLx/+yyP7h2x/v7Ptwdmc7njs4dpxDpFh4TfhXebpY/2T0TX2Y/Pt8nR9s1KDBtOw5WZgFpGFwlq1cdif3O1RiZpUzsz2PbRPCPTch6+7EcKeAZw8Az7Z62h7gJ2AQN0cB8rGK4K7VwAj7fmFvJGmD4Ncn7OXwTaKPZOYWJeALqUTii1wqV0KldK6fLYvhQ3N8gOI9XXyAL9mUaDTCezKoB6+2cP60OpxXGn/ENwv8xsi+Vh4MDi77e/O6cazyUZ4KcCnPEOGFRLCsmoKvrSSfbrHI98Yr3N5E/c+M+rf7WLbuX476bSif39KjeQEo53vF0i0U/OUzQhMB1lGErXe9JCMAqaEtAkwDPhMB6oAXCPBBeHVFgHa5ZQMUdIuPfiMB7XKk7tNBX3RYLgltSXJGKRwg5lkyRBE2lQKtJAFSiHZJ0ith2mpHwM63BNJDcEsgV+5XdCpW+wstg7QhsNszqIwDH98hq7T80ZZAGuy13xIAAXYPC/DJ/w/g1MtfCcDOQfZNPqepJyAJN1/TD5CDb9knyGF2WjwCm4EeRVFEvB3orRr8aZYaIBanh2LG3svOBwAju3M2JDK9nyJxz2I4lCFB7lFvygWNlAtOnWK1wAX2eTknULXA0ZY1lOvGcZfIk1TM7dyFigaFjb0mGqwVDQ5qc6a6agBSRFvVwCmXEUsom6rBc+EFqgYgvLr8P40USgmhEElGCxJwVPD3LslGKlH8cxzhcBWvvVKIMWPUSIMu7gDVBG3SAIeZUEFRkodEw2W0OTK8OAgvgAJCy4+MKl6gFSJUruYtZWIVFRs+iSNs2tl5UyhhRzWHiKlB6aMQ1L1sWVsO8g72i1LY7eAnSWF73ZopbLqRsP0cFnzPcfB6mPFY9SplRkaGbwUuwMxosOrVyAuoEFFe3xuoPSBRqShtmFdSdeyFOHmr9h6l8Wm/HKBa0EMkDWSbp0Q5u5HVjckyREk6eJYkMiaA2B9A9J7cxIK2P7cag/agGNSULRqCF9rNrKlsAUYk5Sr2SEqmJzy5O0rU2eh6M7ruDIpF6z6g6r0u5NqpCzaOftVe58p3G8yuhWpF2OtfL9F7mBS68sWKl53QAxISGvimlXFwigDPjJYpYjoZ+tAFXoAB0dUVEdhV7z2UGxnRX0u8xEePPSHU40QVIJOStZGERkkDVKrblQS7uv3lk0gw5g/88JQWmGGJBpZAxeiWaVLV0PAx/5IEGLGmjNgy4GnLwtQVNFDBqfmY0RdEwCUlWb33AyMVh+aH2zu0VMC/nCWE4jLESYjxpManyWPbJRDwsx0tEwj+3YZILEbCnuHmJiQsJHzzxBAWnh0ZLumqiVgthrfV/bgCldBYLkFcJ5UMMI0wsGBawhiooVY3wqxiG+wEaoOd9ss0aOSdYJAHcG3MtMGKCvAI6k/peO/pgkHIaxMA6J3QX7Pi1Ty4UA8MAlfXj6OX3/NPctIZii7kXhez/fIgtIB+0bFVn68qhMYP2COTfeqDHvrxn4agF8PsP8WJj+X+ayHn4j8=</diagram></mxfile>
+<mxfile host="Electron" modified="2021-09-01T20:06:57.883Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/14.9.6 Chrome/89.0.4389.128 Electron/12.0.16 Safari/537.36" etag="CvbTsLEWPp9N0RejcIko" version="14.9.6" type="device"><diagram id="9mXvaVSuT-yr561JVfZQ" name="Page-1">7V1dd9o4EP01nJM+pMcf4DiPgXy03WSbNm2S7sseYQSoMRZrCwL99SvZMrbxYEhimbTROXlAsgxY987oamZEWnZvsrgI0XR8RQfYb1nGYNGyT1uWZbYtqyX+jMEy6XHTjlFIBnJQ1nFDfmHZacjeGRngqDCQUeozMi12ejQIsMcKfSgM6WNx2JD6xU+dohEuddx4yC/33pEBG8un6BhZ/wdMRuP0k01DXpmgdLDsiMZoQB9zXfZZy+6FlLLk1WTRw76YvHRekvvON1xdfbEQB2yXG26H/cs5+Xn5/c6Z/Ys+ud/vPy8O02mO2DJ9YjzgEyCbNGRjOqIB8s+y3m5IZ8EAi7c1eCsbc0nplHeavPMnZmwp0UQzRnnXmE18eTViKGQnAh3eEdAAp33nxPfl2+JgkI7o+9R7SLpyA8rPL6ckorPQkw8T3vrnk3+OPi7bf308v/gwsG7OxoeSgfwDR5hVTU4yTkxI7gPk7F5gOsEsXPIBIfYRI/MiY5Ak3mg1LsOGv5DwwFBVfes58mfyk65wFAn6WsbHgOEQeYzQoIRn9EgmPorneEgDlkIrcEA+GQX8tYfF/bxjjkNGOPlP5AUmAO16Y+IPLtGSzsRscaC8h7TVHdOQ/OJviwrYSuQtpzDiRtwp0QtxxMdcp+iZa11XaFEYeIkiJjs86vtoGpH+6jEmHEgSdCljdLKNHOIJ8aISTnnVsqVtSM917EgEHjM/YDpyzDjnA+y2Iga0Swwoge0TaUwhfVg5LDFHQ244PerTMLO4FH0fDxmA/YQMBrG5R1PkkWD0LTbuQzPruYxvPLWznq9yFuzYSzDEUIKSgMRHfexf04jELLVPw2Rsd0pJwOKp6nRbndO4J2Q9GvCHQCSGD3P0H7FgAABspbFsR1uiy6m6E7juy7GFfY1ZwvLPd8TOb+mInZIZXo/F1Gk3rMINt1+bF3a1F97VCzs7g928Fwa/sFkWWScj8YzamuuxZtcsmrObiqytiDuqILdLkB8KrGdTHM5JJEz1hLdvsvY6F/ijs5W1r1n30w0+YUfRtttl2xZdlN879OOVeczdBA4Aey/adZfPbM943xEWbvV428zadRn9yoZqt3pLGQe0slYIb2fHFVuZU+8A6HZFbAmzzKwP3hUMXVt53TRwdvT0dVh5/z8aWb9G1tdb6yEc31q257iHtr2PLRZeEHYvbufzn7R+yDcTr08X+cZSNva8LTOBfRm8y3Zf1cbMLO/MKpZtLeGeJ+FMw1jTcMeAZbuAZZttZQv4ESjiJihAIywV3JVsaMde7dhr2bRB8Ktz7GX5JtBHYucWJeBzV4n4Bzm+mAm5pXNG2bRoPtTHB0jvqeID/JV1iEYhvEfubvAqk/PHm+W89PEHbDnF77TbV8oDd+9uv3Jf14+9fJSnAhzK00R4IRFMc0eHrywkn5ZY5HPjG8xeq/5nqv5VHcvK/Muq34L28yt61O8Ayvu9YugWEn/5HaFWgLt4hJV1vWRHAFJDmQJMBZ9WgCrgBQQ+CK8qBWiVUzZAQLe49GsX0CxHdl0d1KnDckhoRZIT34cFYp4lXRRhHSlQShJgC9EsSdolTBvNCFj5lEB6CU4J5ML9kk7FaH8hZZAmBNZzBht14PaSWenLt6YEUrHXfEoABNjZL8BHfx7AqZW/EoDtvdRNPiepxyEJl/fpG4jGj+wdRDO7LW6ByUDPR1FEvDXozR34Uy81QCyO98WMyq+dFwA9q3XSJWJ7P0T8mXmzKyRBbqnX4YJawgXHa8XzDlDnZR9B0QKnhiARuutbR9efeu7FzZfZ9cU9xj3jEK7zmqBFXL2bK92VUQPy5mOFJbh3EYFupeIDIgMgDdRFBsrZA0kDeZpGE+E50t/dwJinxBCUEQF0B3DOWPOgYYcAZBObdQh2OXdQQlmHCp9r5qbR3g3fGmKF8P4AigQLzUei7ixaplWfn/sRDufxvMrwT59SX0d+lPECyBk2bPebiIHmiPhiOm98yqdR0uEbv0KHrbUzXgk9NpOI6OihQg7ZDa4dsHPZy+n5FwUfVo3fJPiQHv/bGnxIS0Cbjz6AR1bd18OMbXHHlBkZGX4UuAAzY8/xSogXv8HZ4TagNjcEpfRhYhVVLEfvZdVYGpg6KgemTDAwVUcZC8wJs8SJQ3nPSQx7yBdwLR+q5UP7yclHqGy90QgEfGZBpKMT4HMO4JvofNsMeE7soW1UMgAqVG+UAXD92mAWomQ3r6NOipCHatgb3X22od2njjrV5NqhqBOIr6qoUzn12BNqyeOmbPQSYaYlXT2SznaLmcZOpyzo2gZk28c1ZBqr8uDlyNKmA2m61GyzS6i0r5eIPZgUqpb6sr+Pjyg+Iu5Cg1E52VTgBxBzfOMM2aIJNpPmKWJQGUMqA1jrFPFEVGJjUvIKTUGW9Fqr806aMEpcCqAhm3UpcNgg4cuGmhbNln25F0iSNuxfdKJbnTMwgfo2EF5VOw5r02FYLsypPy/4gi8zPMMH2xSolKsyt5mkw9+0U1DBmgbXEJg25UUkpc2IRJwyf+PHHHUOCquGQbKWpolCmjj7XjvSpQI6THWXbGFip9Kjs4Cl5RA6cqmACsBvLTUsIzbIzijimGhXse84h9lgoAN2FXCdPXcU5yFONMaTiqp0oKxZAqWZif0RCK7MjvhkJOzpLq9DQkPClk/UsPDoSHNJWdC1QXlbVfyToxLqiymIEzGCAbrIRkGRjVkssXGhbNxxB9gbq6IB9HsROvJRaTEvKqMBwVWEbbmGItmZjFF0JqopdYX/Xmhh72jzylz/pnhY7GYP9B5EHfTQD33XBD1vZv8wL76W+7eD9tn/</diagram></mxfile>
