+++ libs/interactions/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/interactions.module';
 export * from './lib/interfaces';
-export * from './lib/classes';
+export * from './lib/models';
+export * from './lib/enums';
