--- libs/interactions/src/lib/classes/message-interaction.class.ts
@@ -1,5 +1,0 @@
-import { BaseInteraction } from "./base-interaction.class";
-
-export class MessageInteraction extends BaseInteraction {
-  public duration = 5;
-}
