--- libs/interactions/src/lib/classes/phone-interaction.class.ts
@@ -1,5 +1,0 @@
-import { BaseInteraction } from "./base-interaction.class";
-
-export class PhoneInteraction extends BaseInteraction {
-  public duration = 10;
-}
