+++ libs/interactions/src/lib/enums/index.ts
@@ -1,0 +1,1 @@
+export * from './interaction-types.enum';
