+++ libs/interactions/src/lib/enums/interaction-types.enum.ts
@@ -1,0 +1,4 @@
+export enum InteractionTypes {
+  PhoneInteraction,
+  MessageInteraction
+}
